+++ BudgetRegistry/Model/SpendingItemModel.cs
@@ -11,6 +11,6 @@ namespace BudgetRegistry.Model
         public int Id { get; set; }
         public string Name { get; set; }
         public int CategoryId { get; set; }
-        public int LastValue { get; set; }
+        public long LastValue { get; set; }
     }
 }
